chore: add integration tests
