+++ services/api-gateway/internal/connectors/openai/stream_test.go
@@ -22,7 +22,7 @@ func TestRequestStream(t *testing.T) {
 	requestConfigurationDTO := &dto.RequestConfigurationDTO{
 		ApplicationIDString: db.UUIDToString(&application.ID),
 		ApplicationID:       application.ID,
-		PromptConfigID:      &promptConfig.ID,
+		PromptConfigID:      promptConfig.ID,
 		PromptConfigData: datatypes.PromptConfigDTO{
 			ID:                        db.UUIDToString(&promptConfig.ID),
 			Name:                      promptConfig.Name,
